Feat/cant delete old appointments (#844)
* chore(lint): Appointments.jsx

* feat(appointments): do not display delete buttons for appointments before beginning of last month

+++ frontend/src/components/Psychologist/Appointments/Appointments.jsx
@@ -1,11 +1,23 @@
 import React, { useEffect, useState } from 'react';
 import { HashLink } from 'react-router-hash-link';
-import { Button, Table, Icon, TextInput, Callout, CalloutText } from '@dataesr/react-dsfr';
+import {
+  Button,
+  Table,
+  Icon,
+  TextInput,
+  Callout,
+  CalloutText,
+} from '@dataesr/react-dsfr';
 
 import MonthPicker from 'components/Date/MonthPicker';
 
 import agent from 'services/agent';
-import { formatFrenchDate, formatMonth, utcDate } from 'services/date';
+import {
+  formatFrenchDate,
+  formatMonth,
+  getFirstDayOfLastMonth,
+  utcDate,
+} from 'services/date';
 import Badges from 'components/Badges/Badges';
 import { useStore } from 'stores/';
 import getBadgeInfos from 'src/utils/badges';
@@ -13,7 +25,9 @@ import { useNavigate } from 'react-router-dom';
 import styles from './appointments.cssmodule.scss';
 
 const Appointments = () => {
-  const { commonStore: { setNotification } } = useStore();
+  const {
+    commonStore: { setNotification },
+  } = useStore();
 
   const [appointments, setAppointments] = useState([]);
   const [filteredAppointments, setfilteredAppointments] = useState([]);
@@ -24,42 +38,50 @@ const Appointments = () => {
     month: new Date().getMonth() + 1,
   });
 
+  const cantDeleteBefore = getFirstDayOfLastMonth();
+
   const navigate = useNavigate();
   const badges = getBadgeInfos();
 
   useEffect(() => {
-    agent.Appointment.get({ month: month.month, year: month.year })
-      .then(response => {
+    agent.Appointment.get({ month: month.month, year: month.year }).then(
+      (response) => {
         setAppointments(response);
-      });
+      },
+    );
   }, [month]);
 
   useEffect(() => {
     setFilteredMonthApointments(getFilteredMonthAppointments());
   }, [appointments, filteredAppointments]);
 
-  const deleteAppointment = appointmentId => {
+  const deleteAppointment = (appointmentId) => {
     setNotification({});
-    agent.Appointment.delete(appointmentId).then(response => {
+    agent.Appointment.delete(appointmentId).then((response) => {
       setNotification(response);
-      setAppointments(appointments.filter(appointment => appointment.id !== appointmentId));
+      setAppointments(
+        appointments.filter((appointment) => appointment.id !== appointmentId),
+      );
     });
   };
 
-  const handleSearch = e => {
+  const handleSearch = (e) => {
     e.preventDefault();
     e.stopPropagation();
-    const newFilteredAppointments = appointments.filter(patient => patient.lastName.toLowerCase().includes(e.target.value.toLowerCase())
-      || patient.firstNames.toLowerCase().includes(e.target.value.toLowerCase()));
+    const newFilteredAppointments = appointments.filter(
+      (patient) =>
+        patient.lastName.toLowerCase().includes(e.target.value.toLowerCase()) ||
+        patient.firstNames.toLowerCase().includes(e.target.value.toLowerCase()),
+    );
     setfilteredAppointments(newFilteredAppointments);
   };
 
-  const handleSearchClick = e => {
+  const handleSearchClick = (e) => {
     e.preventDefault();
     e.stopPropagation();
   };
 
-  const handleIconClick = id => {
+  const handleIconClick = (id) => {
     if (showTooltip === id) {
       setShowTooltip(null);
     } else {
@@ -68,11 +90,14 @@ const Appointments = () => {
   };
 
   const getFilteredMonthAppointments = () => {
-    const showAppointments = filteredAppointments.length > 0 ? filteredAppointments : appointments;
-    return showAppointments.filter(appointment => {
+    const showAppointments =
+      filteredAppointments.length > 0 ? filteredAppointments : appointments;
+    return showAppointments.filter((appointment) => {
       const appointmentDate = utcDate(appointment.appointmentDate);
-      return appointmentDate.getFullYear() === month.year
-        && appointmentDate.getMonth() === month.month - 1;
+      return (
+        appointmentDate.getFullYear() === month.year &&
+        appointmentDate.getMonth() === month.month - 1
+      );
     });
   };
 
@@ -93,29 +118,28 @@ const Appointments = () => {
     {
       name: 'date',
       label: 'Date',
-      render: appointment => (
+      render: (appointment) => (
         <div className={styles.date}>
           <div>{formatFrenchDate(utcDate(appointment.appointmentDate))}</div>
-          {(appointment.badges.includes(badges.switch_rule_notice.key) || appointment.badges.includes(badges.inactive.key))
-            && (
-              <div
-                className={styles.clickableElement}
-                onClick={() => handleIconClick(appointment.id)}
-              >
-                <Icon
-                  name="ri-information-line"
-                  size="lg"
-                  color="#000091"
-                  iconPosition="right"
-
-                />
-                {showTooltip === appointment.id && (
-                  <span className={styles.tooltip}>
-                    {badges.switch_rule_notice.tooltip}
-                  </span>
-                )}
-              </div>
-            )}
+          {(appointment.badges.includes(badges.switch_rule_notice.key) ||
+            appointment.badges.includes(badges.inactive.key)) && (
+            <div
+              className={styles.clickableElement}
+              onClick={() => handleIconClick(appointment.id)}
+            >
+              <Icon
+                name="ri-information-line"
+                size="lg"
+                color="#000091"
+                iconPosition="right"
+              />
+              {showTooltip === appointment.id && (
+                <span className={styles.tooltip}>
+                  {badges.switch_rule_notice.tooltip}
+                </span>
+              )}
+            </div>
+          )}
         </div>
       ),
       sortable: true,
@@ -124,9 +148,17 @@ const Appointments = () => {
     {
       name: 'badge',
       label: '',
-      render: appointment => {
-        const isInactive = appointment.badges && appointment.badges.includes(badges.inactive.key);
-        return <Badges isInactive={isInactive} badges={appointment.badges} univYear={appointment.univYear} />;
+      render: (appointment) => {
+        const isInactive =
+          appointment.badges &&
+          appointment.badges.includes(badges.inactive.key);
+        return (
+          <Badges
+            isInactive={isInactive}
+            badges={appointment.badges}
+            univYear={appointment.univYear}
+          />
+        );
       },
     },
     {
@@ -143,12 +175,18 @@ const Appointments = () => {
         </div>
       ),
       render: ({ patientId, firstNames, lastName }) => (
-        <div data-test-id="etudiant-name" className={styles.hoverElement} onClick={() => navigate(`/psychologue/modifier-etudiant/${patientId}/#anchor-student-list`)}>
+        <div
+          data-test-id="etudiant-name"
+          className={styles.hoverElement}
+          onClick={() =>
+            navigate(
+              `/psychologue/modifier-etudiant/${patientId}/#anchor-student-list`,
+            )
+          }
+        >
           <span className={styles.tooltip}>Séances de l&apos;étudiant</span>
           <div>
-            {firstNames}
-            {' '}
-            {lastName}
+            {firstNames} {lastName}
           </div>
         </div>
       ),
@@ -156,29 +194,36 @@ const Appointments = () => {
     {
       name: 'actions',
       label: '',
-      render: appointment => (
-        <>
-          <Button
-            data-test-id="delete-appointment-button-small"
-            onClick={() => deleteAppointment(appointment.id)}
-            secondary
-            size="sm"
-            icon="ri-delete-bin-line"
-            className="fr-unhidden fr-hidden-sm fr-float-right"
-            aria-label="Supprimer"
-          />
-          <Button
-            data-test-id="delete-appointment-button-large"
-            secondary
-            size="sm"
-            onClick={() => deleteAppointment(appointment.id)}
-            icon="ri-delete-bin-line"
-            className="fr-hidden fr-unhidden-sm fr-float-right"
-          >
-            Supprimer
-          </Button>
-        </>
-      ),
+      render: (appointment) => {
+        const canDelete =
+          Date.parse(appointment.appointmentDate) >= cantDeleteBefore;
+        if (!canDelete) {
+          return null;
+        }
+        return (
+          <>
+            <Button
+              data-test-id="delete-appointment-button-small"
+              onClick={() => deleteAppointment(appointment.id)}
+              secondary
+              size="sm"
+              icon="ri-delete-bin-line"
+              className="fr-unhidden fr-hidden-sm fr-float-right"
+              aria-label="Supprimer"
+            />
+            <Button
+              data-test-id="delete-appointment-button-large"
+              secondary
+              size="sm"
+              onClick={() => deleteAppointment(appointment.id)}
+              icon="ri-delete-bin-line"
+              className="fr-hidden fr-unhidden-sm fr-float-right"
+            >
+              Supprimer
+            </Button>
+          </>
+        );
+      },
     },
   ];
 
@@ -191,23 +236,19 @@ const Appointments = () => {
           </span>
           <ul>
             <li>
-              <b>Depuis le 1er juillet 2024, 12 séances maximum</b>
-              {' '}
-              prises en charge par année universitaire, sans lettre d&apos;orientation, au tarif de 50€.
+              <b>Depuis le 1er juillet 2024, 12 séances maximum</b> prises en
+              charge par année universitaire, sans lettre d&apos;orientation, au
+              tarif de 50€.
             </li>
             <li>
-              Année universitaire : du
-              {' '}
-              <b>1er septembre au 31 août</b>
-              {' '}
-              de l&apos;année suivante
+              Année universitaire : du <b>1er septembre au 31 août</b> de
+              l&apos;année suivante
             </li>
             <li>
-              Nous vous rappelons que dans le cadre du dispositif, il est strictement
-              {' '}
-              <b>interdit de demander aux étudiants une avance de frais</b>
-              {' '}
-              ou un complément.
+              Nous vous rappelons que dans le cadre du dispositif, il est
+              strictement{' '}
+              <b>interdit de demander aux étudiants une avance de frais</b> ou
+              un complément.
             </li>
           </ul>
         </CalloutText>
@@ -225,9 +266,14 @@ const Appointments = () => {
         </HashLink>
       </div>
       <div className="fr-mb-2w">
-        Veuillez trouver ci-dessous vos séances déclarées pour le mois sélectionné :
+        Veuillez trouver ci-dessous vos séances déclarées pour le mois
+        sélectionné :
         <div className="fr-mt-1w">
-          <MonthPicker month={month} setMonth={setMonth} id="appointment-month" />
+          <MonthPicker
+            month={month}
+            setMonth={setMonth}
+            id="appointment-month"
+          />
         </div>
       </div>
       <div id="appointments-table">
@@ -241,10 +287,8 @@ const Appointments = () => {
           />
         ) : (
           <div>
-            Vous n‘avez pas déclaré de séances pour le mois de
-            {' '}
-            {formatMonth(month)}
-            .
+            Vous n‘avez pas déclaré de séances pour le mois de{' '}
+            {formatMonth(month)}.
           </div>
         )}
       </div>
